+++ CHANGELOG.md
@@ -9,6 +9,11 @@ CHANGELOG
   asking interactively as the final option.
   [#5327](https://github.com/pulumi/pulumi/pull/5327)
 
+- feat(autoapi): Add support for working with private Git repos. Either `SSHPrivateKeyPath`,
+  `PersonalAccessToken` or `UserName` and `Password` can be pushed to the `auto.GitRepo` struct
+  when interacting with a private repo
+  [#5333](https://github.com/pulumi/pulumi/pull/5333)
+
 ## 2.10.0 (2020-09-10)
 
 - feat(autoapi): add Upsert methods for stacks